requirements.txt
Flask
python-dotenv
google-generativeai
PyPDF2
python-docx
werkzeug